+++ backend/plugins/_template/__init__.py
@@ -1,0 +1,1 @@
+"""Copyable plugin template. This package is not enabled by default."""